Extensions.Monad

Description

Documentation

takeJust :: Maybe a -> a

Take Just value or throw an error.

rightsIfAll :: [Either String b] -> Either String [b]

Convert table of Eithers to one Either with table of values.

The result will be Left if there is at least one Left in the initial table. String from the Left value will provide all the Lefts values found during mapping.

(<$$>) :: (Functor f, Functor g) => (a -> b) -> f (g a) -> f (g b)

It is the same as writting:

 fmap fmap fmap
or
 (f $) $ g

(<.>) :: Applicative f => f (a -> b) -> a -> f b infixl 4

It is a helper to avoid writing:

>>> f <*> pure a

ifdM :: Monad a => a () -> Bool -> a ()

Returns a () if True, or return () if False.

ifdrM :: Monad a => a () -> Bool -> a ()

Returns a () if False, or return () if True.

module Control.Monad