Extensions.Parsec

Description

Documentation

module Text.ParserCombinators.Parsec

manyTill1 :: Stream s m t => ParsecT s u m a -> ParsecT s u m end -> ParsecT s u m [a]

The same as manyTill but at least one a is parsed.

spaceb :: CharParser () Char

Very similar to space but without new lines charactes.

spacebs :: CharParser () String

Very similar to spaces but without new lines charactes.

spacebs' :: CharParser () ()

The same as spacebs but discards any results.

eol :: CharParser () String

Read the end of line.

eol' :: CharParser () ()

As eol but discards any results.

spacebsAndEol' :: CharParser () ()

Parse all spacesbs until the end of line that it also read.

btwnSpacebs :: CharParser () a -> CharParser () a

Parse between spacebs. Spacebs are discarded and only the result of the given parser is returned.

letters1 :: CharParser () String

>>> 'many1' 'letter'

spacebs1 :: CharParser () String

>>> 'many1' 'spaceb'

choiceString :: [String] -> CharParser () String

Parsing with trying to use one of the given strings. Very similar to choice.

stringSpacebs :: CharParser () String

Parse a string with spacebs between. End parsing when there is no more letters after spacebs or the charactes is different than letter.

sQuotaBtwnSpacebs :: CharParser () String

Parse ' and discards spacebs around it.