Game.Command

Description

This module provides Commands data and the mean to parse it from a string.

Documentation

parseCommand :: String -> Either ParseError Command

Parse the given string to a Command.

[] are used to show alternative commands, {} are used to show what a player can write in.

Travel:

 [go to\travel to\move to\gt] {locationName}
 ex. go to Old House

ItemsOnObject:

 [use items\ui] '{item},{item2},...' on {object name}
 ex. ui Axe on New Door

UniqueCommand:

 {command} [on// ] {object name}
 {command}
 ex. paint New Door
 ex. super not useful command

CheckBp:

 [check bp/check backpack/cbp]

ExitGame:

 [:exit//:q]

ExitAndSave:

 [:exitSave//:qs]

Help:

 [:help//:h]

PickUpItem:

 [pick up//pup] {item name}
 ex. pup Fantastic Sword

ThrowItem:

 [throw item//thri] {item name}
 ex. thri Fantastic Sword

data Command

Available commands in a game.

Constructors

Travel LocName

ItemsOnObject (Set Item) Object

UniqueCommand String (Maybe Object)

CheckBp

PickUpItem Item

ThrowItem Item

ExitGame

ExitAndSave

Help