Game.CommandHandler

Description

Documentation

data ExitCode

A data type is used to communicate the result, when exiting the game.

Constructors

QuitGame
A player quits game without saving it.

QuitAndSave
A player quits game with saving it.

PlayerWin
A player won the game.

handleCommand :: Command -> GameStateM (Either ExitCode) String

The core of the game logic. Basing on the received Command from the player it manipulates the GameState.

If the game has been finished, regardles of the reason, it will end up in the Left state with ExitCode.

Otherwise it will just return a string as a result message after handling a particular command.