Game.GameState

Documentation

type GameStateIO a = StateT GameStatus IO a

StateT IO monad with GameStatus

type GameStateM m a = StateT GameStatus m a

StateT monad with GameStatus

getLocations :: Monad m => GameStateM m (LocationP, LocMap)

Helper for taking only locations data (LocationP, LocMap) from GameStatus.

getPlayerStatus :: Monad m => GameStateM m PlayerStatus

Helper for taking only PlayerStatus data from GameStatus.

showDescription :: Monad m => GameStateM m String

Provide the whole description of the Location the player is in, taking Conditions into consideration if necessary. if Condition for description of the current Location is not met it will not be added to the result.

checkCondition :: Monad m => CondKey -> GameStateM m Bool

Check whetever the given Condition, indexed by CondType and CondId, is true in current state of a game.

checkConditionWith :: Monad m => CondKey -> a -> a -> GameStateM m a

The same as with checkCondition but you can provide what values should be returned on result

checkConditionWith (cl, id) trueValue falseValue

checkConditionWithM :: Monad m => CondKey -> GameStateM m a -> GameStateM m a -> GameStateM m a

The same as with checkConditionWith but in Monad version.

checkConditionWithM (cl, id) (Just truValue) Nothing

type GameStatus = (PlayerStatus, WorldStatus)

data PlayerStatus

Information about player

LocationP

Current location of the player.

ItemSet

Items carry on by the player.

Constructors

PlayerStatus LocationP ItemSet

psItemExists :: PlayerStatus -> Item -> Bool

Check whether the player has the item.

type WorldStatus = (GameOptions, LocMap)

data GameOptions

Constructors

GameOptions

Fields

goGameName :: String

goGameVersion :: String

goPlayerCapacity :: Int

Define how many items can player can have.

goEndingLocation :: LocName

Define in which location the game will end.

goConditions :: Conditions

Global conditions. They apply to every location in the game.

goIntro :: String

Introduction text when a player starts a new game.

goOutro :: String

Outro text when a player finishes a game.

goCondition :: GameOptions -> CondId -> Condition

Get a global condition from the game options by its Id.

data Location

Data providing all necessary information about location.

Constructors

Location

Fields

lcDescList :: DescMap

lcTravelList :: TravelMap

lcObjects :: ObjectSet

lcItems :: ItemSet

lcActions :: Actions

lcConditions :: Conditions

newtype LocName

LocName is the key of the Location data

Constructors

LocName String

type LocationP = (LocName, Location)

type LocMap = Map LocName Location

type DescMap = Map DescOrder LocDesc

type TravelMap = Map LocName LocCanTravel

lcAddItems :: [Item] -> LocationP -> LocationP

Adding items to the given location.

lcRemoveObjects :: [Object] -> LocationP -> LocationP

Removing objects from the given location.

lcRemoveItem :: Item -> LocationP -> LocationP

Removing an item from the given location.

lcItemExists :: LocationP -> Item -> Bool

Check whether the item exists in the given location.

lcObjExists :: LocationP -> Object -> Bool

Check whether the object exists in the given location.

lcCondition :: LocationP -> CondId -> Condition

Take a condition from the location by its Id. Throws exception when there is no such condition.

data LocDesc

Data type that has two constructors: a simple description LocDesc or description displayed only under some condition LocDescCond.

It is a data type for a Location description.

Constructors

LocDesc String

LocDescCond CondType CondId String

newtype DescOrder

Unique key for ordering the description on the screen.

Constructors

DescOrder Int

data LocCanTravel

Data type used to determine whether a player can travel to a location availible in a current location.

LocCanTravel - there is no restraints to travel. LocCannotTravel - there is some condition to be met before you can travel.

Constructors

LocCannotTravel CondType CondId String

LocCanTravel

data Condition

The Condition type stores information about all conditions that need to be satisfied.

Can have empty parameters.

Constructors

Condition ObjectsNotExist ItemsInLocation PlayerHasItems

data CondType

the Enum data to determine whether condition apply only in a location or it is a global condition that can be applied everywhere during the game.

Constructors

CondLocal

CondGlobal

newtype CondId

An unique Id of a condition. Uniquness works separately for each location and for globals ones; in other words two location can have conditions with the same Ids.

Constructors

CondId Int

type CondKey = (CondType, CondId)

type Conditions = Map CondId Condition

type ObjectsNotExist = [Object]

type ItemsInLocation = [Item]

type PlayerHasItems = [Item]

data Action

Action data.

Constructors

ActionUseItemsOnObject
Action of using items on some object.

Fields

aItemsUsed :: ItemSet

aObject :: Object

aComment :: String

aResults :: [ActionResult]

ActionUnique
Action of doing unique command on some object or without an object.

Fields

auCommands :: [String]

aObjectM :: Maybe Object

aComment :: String

aResults :: [ActionResult]

type Actions = [Action]

data ActionResult

Information about results of some Action.

ArAddLocationItems

- After an action items will be added to a location.

ArRemoveObjects

- After an action objects will be removed from a location.

Constructors

ArAddLocationItems [Item]

ArRemoveObjects [Object]

isUniqeAction :: String -> Maybe Object -> Action -> Bool

Check whether the given action is an UniqueAction.

isItemsOnObjectAction :: Set Item -> Object -> Action -> Bool

Check whether the given action is an ItemsOnObjectAction.

newtype Item

Constructors

Item String

newtype Object

Constructors

Object String

type ItemSet = Set Item

type ObjectSet = Set Object

itemSetFromList :: [String] -> ItemSet