Parser.Text.Language

Description

Documentation

sectionMany :: String -> CharParser () a -> CharParser () [a]

Parse many elements in a section: <SectionName \n {many parser} \n />.

Examples

Expand

 sectionMany "TokenName" (spaces *> string "bb" <* spaces) will parse:
 <TokenName
     bb bb bb 
     bb
 />
and returns ["bb", "bb", "bb", "bb"]

sectionManyWithKey :: String -> CharParser () a -> CharParser () key -> CharParser () (key, [a])

Parse many elements in a section with key: <SectionName n {key parser} {many parser} />.

Examples

Expand

 sectionManyWithKey "TokenName" (spaces *> string "bb" <* spaces) (string "key=1n") will parse:
 <TokenName
     key=1
     bb bb bb 
     bb
 />
and returns ("key=1n", ["bb", "bb", "bb", "bb"])

sectionWithKey :: String -> CharParser () a -> CharParser () key -> CharParser () (key, a)

Parse element in a section with key: <SectionName n {key parser} {parser} />.

Examples

Expand

 sectionWithKey "TokenName" (spaces *> string "bb" <* spaces) (string "key=1n") will parse:
 <TokenName
     key=1
     bb
 />
and returns ("key=1n", "bb")

section :: String -> CharParser () a -> CharParser () a

Parse element in a section <SectionName n {parser} />.

Examples

Expand

 section "TokenName" (spaces *> string "bb" <* spaces) will parse:
 <TokenName
     bb
 />
and returns "bb"

inlineSection :: String -> CharParser () [String]

Parse element in an inline section SectionName \n {parser} />.

Examples

Expand

 section "TokenName" (spaces *> string "bb" <* spaces) will parse:
 <TokenName
     bb
 />
and returns "bb"

beginToken :: String -> CharParser () ()

Parse a begin token: <TokenName, even if surrounded by spaces. It must end with the end-of-line string. It discards result.

endToken :: CharParser () String

Parse the end token: />. It discards the spaces before it.

tryEndToken :: CharParser () String

tryEndToken = try endToken

skipOptionName :: String -> CharParser () ()

Parse an option name and discards results. OptionName = .

Examples

Expand

 skipOptionName "TokenName" will parse:
 TokenName    =    

readOption :: String -> CharParser () String

Parse an option value. OptionName = Value. It must end with the end-of-line string.

Examples

Expand

 readOption "TokenName" will parse:
 TokenName    = "Value"
 and returns "Value"   

quotaString :: CharParser () String

Parse a string between quotas. Quatas in the string are not allowed.

Examples

Expand

 quotaString will parse:
 "Value S"
 and returns "Value S"   

readValueName :: CharParser () String

The same as with quataString but discards preceding spaces and the stream must end with the eol string.

readInt :: CharParser () Int

Read a number. For non-number returns 0.

readOptionInt :: String -> CharParser () Int

the same as with readOption but parses number instead.

tryOptionInt :: String -> CharParser () Int

tryToTV :: (b -> c) -> a -> CharParser () b -> CharParser () (a, c)

Try parsing b and transform it to c and then put it to tupple (a, c).

choiceToTV :: [(a, CharParser () c)] -> CharParser () (a, c)

Try parsing choice of tryToTVs.

string' :: String -> CharParser () ()

module Extensions.Parsec